Added tests for black list feature.
In the process of adding a new feature which will convert the white list into a black list.  The black list feature should be able to work on all current methods.  I want to add a third parameter to all methods which will take a boolean value known as black list.  If set to true, the white list paramater will change into a black list and will only remove tags and entities that are specified.